+++ src/constants/api.js
@@ -59,7 +59,7 @@ export const API_DOMAINS = {
   ADMIN_BOARD_DETAIL: '/admin/boards/:boardId',
   ADMIN_ACTIVE_BOARD: '/admin/boards/:boardId/active',
   ADMIN_INACTIVE_BOARD: '/admin/boards/:boardId/inactive',
-  
+
   ADMIN_CARDNEWS: '/admin/cardNews',
   ADMIN_CARDNEWS_DETAIL: '/admin/cardNews/:postId',
 
@@ -100,7 +100,6 @@ export const QUERY_KEYS = {
 
   GET_COMMENT_LIST: 'commentList',
   GET_TRANSLATE_POST: 'postTranslated',
-  GET_TRANSLATE_POST_LIST: 'postTranslatedList',
   GET_TRANSLATE_TEXT: 'textTranslated',
 
   GET_SEARCH_KEYWORD: 'searchKeyword',
